school/requirements.txt
flake8==2.3.0
mccabe==0.3
pep8==1.5.7
pyflakes==0.8.1
pymongo==2.7.2
pyproj==1.9.3
PyYAML==3.11
openregister-entry==0.2.0
requests==2.6.2
